Adding styles and components
