+++ src/components/cards/inventoryCard.js
@@ -1,0 +1,14 @@
+export default function InventoryCard(){
+
+  return (
+    <div className="card">
+      <div className="title-card">
+        <p className="inventory-icon">
+          M
+        </p>
+        Mercurio
+      </div>
+      <p>Cantidad: 3</p>
+    </div>
+  )
+}
